Change IHasFireRate.FireRate from int to float

+++ Assets/Scripts/Towers/Actions/IHasFireRate.cs
@@ -2,6 +2,6 @@
 {
     public interface IHasFireRate
     {
-        int FireRate { get; }
+        float FireRate { get; }
     }
 }
